move the utils module to the root level of the package
